% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.r
\name{num_int}
\alias{num_int}
\title{This function performes numerical integration using monte carlo integration}
\usage{
num_int(fn, ..., a, b, N = 100, ncores = 1, cl = NULL)
}
\arguments{
\item{fn}{A function}

\item{a}{The lower bound}

\item{b}{is The upper bound}

\item{N}{is The sample size}

\item{ncores}{The number of processors to call}

\item{cl}{An object of class cluster}
}
\description{
This function performes numerical integration using monte carlo integration
}
\details{
\code{code}; \emph{emph}; \pkg{graph}; \CRANpkg{graph}
}
